creating wrapper component
